Move BucketList into it's own file
